feat: extract local storage utilities
